Add start of TypeScript migration
