+++ .github/workflows/main.yml
@@ -9,9 +9,9 @@ jobs:
     strategy:
       matrix:
         node-version:
+          - 20
           - 18
           - 16
-          - 14
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-node@v3
