+++ package.json
@@ -7,12 +7,11 @@
 		"cisco-vpn-rdp-connecter": "dist/cli.js"
 	},
 	"files": [
-		"dist",
-		"index.d.ts"
+		"dist"
 	],
 	"scripts": {
-		"build": "babel . --out-dir=dist --ignore node_modules,dist",
-		"dev": "babel . --out-dir dist --watch --ignore node_modules,dist",
+		"build": "tsc",
+		"dev": "tsc --watch",
 		"prepublish": "npm run build",
 		"test": "tsd",
 		"lint": "eslint ."
@@ -31,9 +30,9 @@
 	],
 	"author": "Knut Kirkhorn",
 	"exports": "./dist/index.js",
-	"types": "./index.d.ts",
+	"types": "./dist/index.d.ts",
 	"engines": {
-		"node": ">=14.16"
+		"node": ">=16"
 	},
 	"license": "MIT",
 	"bugs": {
@@ -43,15 +42,14 @@
 	"dependencies": {
 		"cisco-vpn": "^1.0.2",
 		"conf": "^11.0.1",
-		"import-jsx": "^5.0.0",
-		"ink": "^4.1.0",
+		"ink": "^4.2.0",
 		"ink-select-input": "^5.0.0",
 		"ink-spinner": "^5.0.0",
 		"ink-text-input": "^5.0.1",
 		"is-online": "^10.0.0",
 		"is-unicode-supported": "^1.3.0",
 		"log-symbols": "^5.1.0",
-		"meow": "^11.0.0",
+		"meow": "^12.0.1",
 		"ps-list": "^8.1.1",
 		"react": "^18.2.0",
 		"regedit": "^5.1.1",
@@ -59,20 +57,19 @@
 		"xml-js": "^1.6.11"
 	},
 	"devDependencies": {
-		"@babel/cli": "^7.21.0",
-		"@babel/preset-react": "^7.18.6",
-		"eslint": "^8.38.0",
+		"@sindresorhus/tsconfig": "^3.0.1",
+		"@types/react": "^18.2.6",
+		"@typescript-eslint/eslint-plugin": "^5.59.5",
+		"@typescript-eslint/parser": "^5.59.5",
+		"eslint": "^8.40.0",
 		"eslint-config-airbnb": "^19.0.4",
 		"eslint-config-airbnb-base": "^15.0.0",
 		"eslint-plugin-import": "^2.27.5",
 		"eslint-plugin-jsx-a11y": "^6.7.1",
 		"eslint-plugin-react": "^7.32.2",
-		"eslint-plugin-unicorn": "^46.0.0",
-		"tsd": "^0.28.1"
-	},
-	"babel": {
-		"presets": [
-			"@babel/preset-react"
-		]
+		"eslint-plugin-unicorn": "^47.0.0",
+		"ts-node": "^10.9.1",
+		"tsd": "^0.28.1",
+		"typescript": "^5.0.4"
 	}
 }
