+++ test-d/index.test-d.ts
@@ -14,7 +14,7 @@ import {
 	isRdpWindowOpened,
 	openRdpWindow,
 	RdpDefaults
-} from './index.js';
+} from '../source/index.js';
 
 expectType<Promise<void>>(connectToVpn('example-server', '1', 'knut', 'supersecret'));
 expectType<Promise<void>>(openRdpWindow('example-server'));
