+++ tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "@sindresorhus/tsconfig",
+	"compilerOptions": {
+		"outDir": "dist"
+	},
+	"include": ["source"]
+}
